fix: move holiday expansion out of server actions
